Bugfix for #701 and #702 (#764)
* bug fix for issue #702

* add in debug for system util callouts and duplicate call

removed duplicate calling of system util values
added debug calls to debug log when getting system values

* breakout of drive utilities

* removed codesmells

* bugfix for #701

fix for when users remove/delete a database entry, cleans up  any previous jobs

* fix for #701 drive settings error

* added additional catch for #701

---------
