+++ arm/ui/settings/ServerUtil.py
@@ -39,6 +39,7 @@ class ServerUtil():
             self.cpu_util = psutil.cpu_percent()
         except EnvironmentError:
             self.cpu_util = 0
+        app.logger.debug(f"Server CPU Util: {self.cpu_util}")
 
     def get_cpu_temp(self):
         try:
@@ -49,6 +50,7 @@ class ServerUtil():
                 self.cpu_temp = 0
         except EnvironmentError:
             self.cpu_temp = 0
+        app.logger.debug(f"Server CPU Temp:  {self.cpu_temp}")
 
     def get_memory(self):
         try:
@@ -60,6 +62,9 @@ class ServerUtil():
             self.memory_free = 0
             self.memory_used = 0
             self.memory_percent = 0
+        app.logger.debug(f"Server Mem Free:  {self.memory_free}")
+        app.logger.debug(f"Server Mem Used:  {self.memory_used}")
+        app.logger.debug(f"Server Mem Percent:  {self.memory_percent}")
 
     def get_disk_space(self, filepath):
         # Hard drive space
@@ -73,4 +78,6 @@ class ServerUtil():
             app.logger.debug("ARM folders not found")
             flash("There was a problem accessing the ARM folder: "
                   f"'{filepath}'. Please make sure you have setup ARM", "danger")
+        app.logger.debug(f"Server {filepath} Space:  {disk_space}")
+        app.logger.debug(f"Server {filepath} Percent:  {disk_percent}")
         return disk_space, disk_percent
