+++ arm/ui/templates/index.html
@@ -61,6 +61,7 @@
         <h5 class="text-center"><strong>System Information</strong></h5>
     </p>
     {% include 'settings/sysinfo.html' %}
+
     <script type="application/javascript" src="static/js/common.js"></script>
     <script type="application/javascript" src="static/js/jobRefresh.js"></script>
     <script type="application/javascript">
